requirements.txt
ultralytics>=8.0.0
opencv-python
numpy
pyyaml
optuna
pandas
matplotlib
tqdm